moving Encoding Utils
