+++ distribution/lib/Standard/Base/0.0.0-dev/src/System/Output_Stream.enso
@@ -15,7 +15,7 @@ from project.System.Input_Stream import close_stream
 polyglot java import java.io.ByteArrayOutputStream
 polyglot java import java.io.OutputStream as Java_Output_Stream
 polyglot java import org.enso.base.encoding.ReportingStreamEncoder
-polyglot java import org.enso.base.Encoding_Utils
+polyglot java import org.enso.base.encoding.Encoding_Utils
 
 ## PRIVATE
    An output stream, allowing for interactive writing of contents.
