+++ std-bits/base/src/main/java/org/enso/base/encoding/Encoding_Utils.java
@@ -1,4 +1,4 @@
-package org.enso.base;
+package org.enso.base.encoding;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -15,8 +15,8 @@ import java.util.Arrays;
 import java.util.function.BiConsumer;
 import java.util.function.Function;
 import java.util.function.IntFunction;
-import org.enso.base.encoding.ReportingStreamDecoder;
-import org.enso.base.encoding.ReportingStreamEncoder;
+
+import org.enso.base.WithProblems;
 import org.enso.base.text.ResultWithWarnings;
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Value;
